backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo>=4.15.2
python-dotenv==1.0.1
groq>=0.9.0
requests>=2.32.3
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2
joblib>=1.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sendgrid>=6.11.0
razorpay>=1.4.1
pyotp>=2.9.0
cryptography>=42.0.7